Reduce invalidation when using custom loggers
If you define new AbstractLogger types, it ends up invalidating basically
any bit of code with an atsign-warn, atsign-error, etc.
The key offender is `catch_exceptions`, which has a fallback method
which is lowest in the method table and which therefore gets invalidated
by any other specific type.

This avoids the need to specialize that method by defining a new
`UncaughtExceptionLogger` abstract type. For good measure, it provides
specializations for the other defined logger types.